refactor: pr review 반영
