--- src/constants/userData.ts
@@ -1,1 +1,0 @@
-export const githubURl = 'https://github.com/react-sprint/blog-starter/';
